+++ schema/postgres/sqls/pg_PANDA_TABLE.sql
@@ -2356,7 +2356,7 @@ CREATE TABLE mv_jobsactive4_stats (
 	num_of_jobs bigint,
 	vo varchar(32),
 	workqueue_id integer,
-    num_of_cores integer
+    num_of_cores bigint
 ) ;
 COMMENT ON TABLE mv_jobsactive4_stats IS E'Table (was from Materialized view before, but was not reliable) which collects aggregated data on set of attributes(columns). The data is read from the JOBSACTIVE4 table by an Oracle scheduler job. The refresh interval is 2 min';
 COMMENT ON COLUMN mv_jobsactive4_stats.cloud IS E'cloud (associated with Tier 1) where the job is submitted to';
